Refactored some things to a base token class. Will extend off this class if needed (I don't think it will be needed). Took out bulky delimiter declarations and am going to try to use a nebulous delimiter token type
